Move the utilities connected only to Rejections to `Rejections` and add some JavaDoc.
