+++ client/src/test/java/io/spine/core/CommandsShould.java
@@ -31,14 +31,12 @@ import com.google.protobuf.Int64Value;
 import com.google.protobuf.StringValue;
 import com.google.protobuf.Timestamp;
 import io.spine.Identifier;
-import io.spine.base.ThrowableMessage;
 import io.spine.client.ActorRequestFactory;
 import io.spine.client.TestActorRequestFactory;
 import io.spine.core.given.GivenCommandContext;
 import io.spine.core.given.GivenUserId;
 import io.spine.string.Stringifiers;
 import io.spine.time.Durations2;
-import io.spine.time.Time;
 import io.spine.time.ZoneOffset;
 import io.spine.time.ZoneOffsets;
 import io.spine.type.TypeName;
@@ -49,7 +47,6 @@ import java.util.List;
 
 import static com.google.common.collect.Lists.newArrayList;
 import static com.google.protobuf.Descriptors.FileDescriptor;
-import static io.spine.core.Commands.causedByRejection;
 import static io.spine.core.Commands.sameActorAndTenant;
 import static io.spine.core.given.GivenTenantId.newUuid;
 import static io.spine.protobuf.TypeConverter.toMessage;
@@ -64,7 +61,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 /**
- * * Tests for {@linkplain Commands Commands utility class}.
+ * Tests for {@linkplain Commands Commands utility class}.
  *
  * <p>The test suite is located under the "client" module since actor request generation
  * is required. So we want to avoid circular dependencies between "core" and "client" modules.
@@ -250,22 +247,5 @@ public class CommandsShould {
                                                .toUrl();
 
         assertEquals(TypeUrl.of(StringValue.class), typeUrl);
-    }
-
-    @SuppressWarnings({
-            "NewExceptionWithoutArguments" /* No need to have a message for this test. */,
-            "SerializableInnerClassWithNonSerializableOuterClass" /* Does not refer anything. */
-    })
-    @Test
-    public void say_if_RuntimeException_was_called_by_command_rejection() {
-        assertFalse(causedByRejection(new RuntimeException()));
-        final ThrowableMessage throwableMessage = new ThrowableMessage(Time.getCurrentTime()) {
-            private static final long serialVersionUID = 0L;
-        };
-        assertTrue(causedByRejection(new IllegalStateException(throwableMessage)));
-
-        // Check that root cause is analyzed.
-        assertTrue(causedByRejection(
-                new RuntimeException(new IllegalStateException(throwableMessage))));
     }
 }
