+++ server/src/main/java/io/spine/server/procman/ProcessManagerRepository.java
@@ -56,7 +56,7 @@ import javax.annotation.CheckReturnValue;
 import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static io.spine.core.Commands.causedByRejection;
+import static io.spine.core.Rejections.causedByRejection;
 import static io.spine.util.Exceptions.newIllegalStateException;
 
 /**
